+++ lib/src/algorithms/argon2/argon2_32bit.dart
@@ -3,8 +3,8 @@
 
 import 'dart:typed_data';
 
-import 'argon2.dart';
-import 'blake2b.dart';
+import 'common.dart';
+import '../blake2b_32bit.dart';
 
 const int _mask16 = 0xFFFF;
 const int _mask32 = 0xFFFFFFFF;
@@ -28,50 +28,22 @@ const int _address = _input + 256;
 // |            |            |            |            | -> lane p - 1
 // +------------+------------+------------+------------+
 
-class Argon2Internal extends Argon2 {
+class Argon2Internal {
+  final Argon2Context ctx;
   final _blockR = Uint32List(256);
   final _blockT = Uint32List(256);
   final _temp = Uint32List(_address + 256);
 
-  Argon2Internal({
-    required List<int> salt,
-    required int hashLength,
-    required int passes,
-    required int memorySizeKB,
-    required int lanes,
-    required int segments,
-    required int columns,
-    required int blocks,
-    required int slices,
-    required List<int>? key,
-    required List<int>? personalization,
-    required Argon2Type type,
-    required Argon2Version version,
-  }) : super.internal(
-          salt: salt,
-          slices: slices,
-          version: version,
-          type: type,
-          hashLength: hashLength,
-          passes: passes,
-          lanes: lanes,
-          memorySizeKB: memorySizeKB,
-          segments: segments,
-          columns: columns,
-          blocks: blocks,
-          key: key,
-          personalization: personalization,
-        );
+  Argon2Internal(this.ctx);
 
-  @override
-  Argon2HashDigest convert(List<int> password) {
+  Uint8List convert(List<int> password) {
     int i, j, k, p;
     int pass, slice, lane;
     var hash0 = Uint8List(64 + 8);
     var hash0as32 = hash0.buffer.asUint32List();
-    var buffer32 = Uint32List(blocks << 8);
+    var buffer32 = Uint32List(ctx.blocks << 8);
     var buffer = buffer32.buffer.asUint8List();
-    var result = Uint8List(hashLength);
+    var result = Uint8List(ctx.hashLength);
 
     // H_0 Generation (64 + 8 = 72 bytes)
     _initialHash(hash0, password);
@@ -80,7 +52,7 @@ class Argon2Internal extends Argon2 {
     // Lane Starting Blocks
     k = 0;
     hash0as32[16] = 0;
-    for (i = 0; i < lanes; i++, k += columns) {
+    for (i = 0; i < ctx.lanes; i++, k += ctx.columns) {
       // B[i][0] = H'^(1024)(H_0 || LE32(0) || LE32(i))
       hash0as32[17] = i;
       _expandHash(1024, hash0, buffer, k << 10);
@@ -89,16 +61,16 @@ class Argon2Internal extends Argon2 {
     // Second Lane Blocks
     k = 1;
     hash0as32[16] = 1;
-    for (i = 0; i < lanes; i++, k += columns) {
+    for (i = 0; i < ctx.lanes; i++, k += ctx.columns) {
       // B[i][1] = H'^(1024)(H_0 || LE32(1) || LE32(i))
       hash0as32[17] = i;
       _expandHash(1024, hash0, buffer, k << 10);
     }
 
     // Further block generation
-    for (pass = 0; pass < passes; ++pass) {
-      for (slice = 0; slice < slices; ++slice) {
-        for (lane = 0; lane < lanes; ++lane) {
+    for (pass = 0; pass < ctx.passes; ++pass) {
+      for (slice = 0; slice < ctx.slices; ++slice) {
+        for (lane = 0; lane < ctx.lanes; ++lane) {
           _fillSegment(buffer32, pass, slice, lane);
         }
       }
@@ -106,10 +78,10 @@ class Argon2Internal extends Argon2 {
 
     // Finalization
     /* XOR the blocks */
-    j = columns - 1;
+    j = ctx.columns - 1;
     var block = buffer.buffer.asUint8List(j << 10, 1024);
-    for (k = 1; k < lanes; ++k) {
-      j += columns;
+    for (k = 1; k < ctx.lanes; ++k) {
+      j += ctx.columns;
       p = j << 10;
       for (i = 0; i < 1024; ++i, ++p) {
         block[i] ^= buffer[p];
@@ -117,8 +89,8 @@ class Argon2Internal extends Argon2 {
     }
 
     /* Hash the result */
-    _expandHash(hashLength, block, result, 0);
-    return Argon2HashDigest(this, result);
+    _expandHash(ctx.hashLength, block, result, 0);
+    return result;
   }
 
   void _initialHash(Uint8List _hash0, List<int> password) {
@@ -127,23 +99,23 @@ class Argon2Internal extends Argon2 {
     //         LE32(length(S)) || S ||  LE32(length(K)) || K ||
     //         LE32(length(X)) || X)
     var blake2b = Blake2bHash(64);
-    blake2b.addUint32(lanes);
-    blake2b.addUint32(hashLength);
-    blake2b.addUint32(memorySizeKB);
-    blake2b.addUint32(passes);
-    blake2b.addUint32(version.value);
-    blake2b.addUint32(type.index);
+    blake2b.addUint32(ctx.lanes);
+    blake2b.addUint32(ctx.hashLength);
+    blake2b.addUint32(ctx.memorySizeKB);
+    blake2b.addUint32(ctx.passes);
+    blake2b.addUint32(ctx.version.value);
+    blake2b.addUint32(ctx.type.index);
     blake2b.addUint32(password.length);
     blake2b.add(password);
-    blake2b.addUint32(salt.length);
-    blake2b.add(salt);
-    blake2b.addUint32(key?.length ?? 0);
-    if (key != null) {
-      blake2b.add(key!);
+    blake2b.addUint32(ctx.salt.length);
+    blake2b.add(ctx.salt);
+    blake2b.addUint32(ctx.key?.length ?? 0);
+    if (ctx.key != null) {
+      blake2b.add(ctx.key!);
     }
-    blake2b.addUint32(personalization?.length ?? 0);
-    if (personalization != null) {
-      blake2b.add(personalization!);
+    blake2b.addUint32(ctx.personalization?.length ?? 0);
+    if (ctx.personalization != null) {
+      blake2b.add(ctx.personalization!);
     }
 
     var hash = blake2b.digest().bytes;
@@ -209,9 +181,9 @@ class Argon2Internal extends Argon2 {
     int previous, current;
     int i, j, startIndex, rand0, rand1;
 
-    bool dataIndependentAddressing = (type == Argon2Type.argon2i);
-    if (type == Argon2Type.argon2id) {
-      dataIndependentAddressing = (pass == 0) && (slice < midSlice);
+    bool dataIndependentAddressing = (ctx.type == Argon2Type.argon2i);
+    if (ctx.type == Argon2Type.argon2id) {
+      dataIndependentAddressing = (pass == 0) && (slice < ctx.midSlice);
     }
 
     if (dataIndependentAddressing) {
@@ -221,11 +193,11 @@ class Argon2Internal extends Argon2 {
       _temp[_input + 3] = 0;
       _temp[_input + 4] = slice;
       _temp[_input + 5] = 0;
-      _temp[_input + 6] = blocks;
-      _temp[_input + 7] = blocks >>> 32;
-      _temp[_input + 8] = passes;
+      _temp[_input + 6] = ctx.blocks;
+      _temp[_input + 7] = ctx.blocks >>> 32;
+      _temp[_input + 8] = ctx.passes;
       _temp[_input + 9] = 0;
-      _temp[_input + 10] = type.index;
+      _temp[_input + 10] = ctx.type.index;
       _temp[_input + 11] = 0;
       _temp[_input + 12] = 0;
       _temp[_input + 13] = 0;
@@ -242,19 +214,19 @@ class Argon2Internal extends Argon2 {
     }
 
     /* Offset of the current block */
-    current = lane * columns + slice * segments + startIndex;
+    current = lane * ctx.columns + slice * ctx.segments + startIndex;
 
-    if (current % columns == 0) {
+    if (current % ctx.columns == 0) {
       /* Last block in this lane */
-      previous = current + columns - 1;
+      previous = current + ctx.columns - 1;
     } else {
       /* Previous block */
       previous = current - 1;
     }
 
-    for (i = startIndex; i < segments; ++i, ++current, ++previous) {
+    for (i = startIndex; i < ctx.segments; ++i, ++current, ++previous) {
       /* 1.1 Rotating prev_offset if needed */
-      if (current % columns == 1) {
+      if (current % ctx.columns == 1) {
         previous = current - 1;
       }
 
@@ -275,7 +247,7 @@ class Argon2Internal extends Argon2 {
       }
 
       /* 1.2.2 Computing the lane of the reference block */
-      refLane = rand1 % lanes;
+      refLane = rand1 % ctx.lanes;
 
       if (pass == 0 && slice == 0) {
         /* Can not reference other lanes yet */
@@ -297,9 +269,9 @@ class Argon2Internal extends Argon2 {
         buffer,
         next: current << 8,
         prev: previous << 8,
-        ref: (refLane * columns + refIndex) << 8,
+        ref: (refLane * ctx.columns + refIndex) << 8,
         /* 1.2.1 v10 and earlier: overwrite, not XOR */
-        xor: version != Argon2Version.v10 && pass > 0,
+        xor: ctx.version != Argon2Version.v10 && pass > 0,
       );
     }
   }
@@ -398,16 +370,16 @@ class Argon2Internal extends Argon2 {
         area = index - 1; // all but the previous
       } else if (sameLane) {
         // The same lane => add current segment
-        area = slice * segments + index - 1;
+        area = slice * ctx.segments + index - 1;
       } else {
-        area = slice * segments + (index == 0 ? -1 : 0);
+        area = slice * ctx.segments + (index == 0 ? -1 : 0);
       }
     } else {
       // Other passes
       if (sameLane) {
-        area = columns - segments + index - 1;
+        area = ctx.columns - ctx.segments + index - 1;
       } else {
-        area = columns - segments + (index == 0 ? -1 : 0);
+        area = ctx.columns - ctx.segments + (index == 0 ? -1 : 0);
       }
     }
 
@@ -418,12 +390,12 @@ class Argon2Internal extends Argon2 {
 
     /* 1.2.5 Computing starting position */
     start = 0;
-    if (pass != 0 && slice != slices - 1) {
-      start = (slice + 1) * segments;
+    if (pass != 0 && slice != ctx.slices - 1) {
+      start = (slice + 1) * ctx.segments;
     }
 
     /* 1.2.6. Computing absolute position */
-    return (start + pos) % columns;
+    return (start + pos) % ctx.columns;
   }
 
   /// `v[i]++`